Enable python dependency caching in CI
This adds a new environment variable called DPT_VENV_CACHING which can be set to 1 to
enable venv caching.

When DPT_VENV_CACHING is 1 then the following changes are made to the way the project runs locally:

- poetry install is not executed during image build
- POETRY_VIRTUALENVS_CREATE is set to 1
- The django image uses virtualenv from a volume (inside DPT_POETRY_CACHE_DIR)
- poetry install is executed before starting django container (can be skipped with SKIP_INSTALL env variable)
- Adds python virtualenv bin dir to PATH and sets its value to VIRTUAL_ENV variable

Caveat:

The poetry venv path is "dynamic" and can change when base python is updated or when the project path changes
inside the container. Since we need to know the path when building the docker image we are forced to "hard code"
the path into Dockerfile-django. During startup this hard-coded path is compared to the real venv path and if
it does not match the develpoer needs to update the value in Dockerfile-django. A script and a make commmand
is included to simplify this (`make dev-venv-path`).

As a result of this we can cache the python dependencies in CI. Only thing we need to add is CI caching for
the DPT_POETRY_CACHE_DIR folder.

Note: I have enabled this in one of our internal projects already and it seems to work nicely. I also use
DPT_VENV_CACHING locally and have had no problems myself. But I would not make this the default way to run
projects locally.