IN-1976: Add cypress autotests
This adds Cypress autotests to both SPA and webapp styles.

The Cypress code is kept under the `cypress/` directory and then copied
into the respective `app/` or `webapp/` directory.

A few Cypress tests were also added, related to authentication in the regular GUI
and the admin.

For SPA, those tests also cover access to the "restricted" page, but for webapp,
there is no such page.

Some Cypress commands related to login and logout were also added.

The Gitlab CI in projects generated with the cypress option turned on
runs the Cypress tests in CLI/CI mode.

No Cypress tests were added to run in CI for the DPT itself, because it would require
an enormous amount of time to complete such a test, due to Cypress being quite resource-heavy;
even if it's technically doable, the amount of time to verify it would probably exceed
the limits of reason to spend on this task.

In the process of adding Cypress, I tried to minimize the Jinja tags usage inside the code.

Testing
-------

Generate a project, and run
```
make setup
make cypress-test
```

It's also possible to create a project in Gitlab and push the generated project there, to verify
that the pipelines are passing.

Issues
------

I have a very odd issue that in Gitlab CI, one cypress test intermittently fails
regardless of the timeout settings and the fact that it passes locally.