Support multiple envs with env caching
Previously when `poetry env list` returned multiple environments then the dev
django entrypoint file raised an error as the ENV_FOLDER setting did not match
the available env string directly.

Improved the check there to handle the case correctly and not error when the ENV_FOLDER
value exists in the returned list of environments.

Also:

- Update production poetry install command