Replace safety with pip audit.
The tool we currently use for python dependency vulnerability scanning, safety, is only allowed for non-commercial projects (starting from version 2.0.0). The warning message we get from the scan is as follows:

  You are using Safety's free vulnerability database. This data is outdated, limited, and
  licensed for non-commercial use only.
  All commercial projects must sign up and get an API key at https://pyup.io

The best replacement seems to be pip-audit and we have implemented using it as a replacement in various projects internally. Pip-audit is mostly
a drop in replacement with only some tiny changes required in the ignoring of issues.

See ignored-vulnerabilities.txt for more info on how to ignore and document ignores.